move enum KeychainError to the new file
